Move ToolRegistry to tools/ and make logger an optional argument in orchestrator
